Add clearer test assertion and increase cluster size in LE tests
With a "split brain" on 5 instances the larget the number the better,
just to make tests complete more reliably. Otherwise it's "possible"
for a split vote. 7 just makes it less likely. It doesn't matter
though, the algorithm is still working, but we just want the test to
complete in one round, rather than N rounds.